+++ test/test-runner.ts
@@ -76,6 +76,24 @@ test('skip-test', async (t) => {
         'second test not marked as skipped'
     )
 })
+
+test('only-test', async (t) => {
+    const res = await runTest('./only-test.js')
+    const testsuites = res.xml.testsuites
+
+    t.equal(res.exitCode, 0, 'test shall pass')
+    t.equal(testsuites.$.tests, '2')
+    t.equal(testsuites.$.errors, '0')
+    t.equal(testsuites.$.failures, '0')
+    t.equal(testsuites.$.skipped, '1')
+    t.true(
+        testsuites.testsuite[0].testcase[0].skipped,
+        'first test marked as skipped'
+    )
+    t.false(
+        testsuites.testsuite[0].testcase[1].skipped,
+        'second test not marked as skipped'
+    )
 })
 
 test('failure-in-beforeAll', async (t) => {