+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 -   Report skipped tests correctly in junit xml
+-   Regard other tests as skipped when using only
 
 ## [3.5.4] 2020-04-22